% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pde-enrich.R
\name{get_expanded_ranges}
\alias{get_expanded_ranges}
\title{Get vector of expanded ranges for peak widths}
\usage{
get_expanded_ranges(x = 5, expand_range_kb = c(5, 5000), ...)
}
\arguments{
\item{x}{\code{numeric} value representing the step size.}

\item{expand_range_kb}{\code{numeric} range of distances to allow
in kilobases (kb) for \code{expand_kb}.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} vector of distances in kilobases.
}
\description{
Get vector of expanded ranges for peak widths
}
\details{
This function takes a range of distances in kilobases,
and a default initial step size,
and returns a reasonable set of human-readable distances
that spans this range in using roughly log-scale steps.
}
\examples{
opar <- par("mfrow"=c(2,2))
for (x in c(2, 5, 10, 20)) {
   y <- get_expanded_ranges(x);
   plot(y,
      ylim=c(0, 5300),
      pch=20,
      xlab="step",
      main=paste("step size", x))
}
par(opar)

}
