% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pde-enrich.R
\name{peak_feature_ranged_counter}
\alias{peak_feature_ranged_counter}
\title{Count peak-feature overlap with increasing distance}
\usage{
peak_feature_ranged_counter(
  peaksL = NULL,
  featuresL = NULL,
  randomsL = NULL,
  category_colname = "hit",
  default_category = "hit",
  expand_kb = NULL,
  expand_range_kb = c(5, 5000),
  expand_factor = 2,
  do_random = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{peaksL}{list named by experiment, containing a
list of GRanges, where each set of GRanges represents
a set of peaks.}

\item{featuresL}{list named by experiment, containing
GRanges representing features, with \code{category_colname}
containing category values.}

\item{randomsL}{list named by experiment, containing
GRanges from which random peaks should be drawn, using
the same number of peaks as each corresponding entry
in \code{peaksL}.}

\item{category_colname}{character string indicating the
colname present in the GRanges for each \code{featuresL}
element. If the values are factors, the factor
levels should be preserved in order. When the
column \code{category_colname} does not exist, or for
any NA entries, \code{default_category} is used as a
replacement.}

\item{default_category}{vector length=1 used when
\code{category_colname} is not present in
\code{colnames(values(featuresL[[i]]))}.}

\item{expand_kb}{numeric vector of distances in kilobases (kb)
to use when expanding the peaks in \code{peaksL}. When
\code{expand_kb} is \code{NULL} then the default values are used,
within the range defined by \code{expand_range_kb}.}

\item{expand_range_kb}{numeric range of distances to allow
in kilobases (kb) for \code{expand_kb}.}

\item{expand_factor}{numeric value multiplied by the expand width
during \code{GenomicRanges::resize()}, since the resize happens
at peak center and defines a fixed width. For example,
use \code{expand_factor=2} when this distance should represent
distance from the peak in either direction, for which the
peaks are expanded from center by this distance.}

\item{do_random}{\code{logical} indicating whether the test being
performed represents a random permutation.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Count peak-feature overlap with increasing distance
}
\details{
This function counts features overlaps across peaks,
while progressively expanding the peak width.
It also tabulates counts by feature category, using
the feature annotations in \code{category_colname}.
}
