% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pde-enrich.R
\name{plot_ranged_counts}
\alias{plot_ranged_counts}
\title{Plot feature-peak overlap counts by distance}
\usage{
plot_ranged_counts(
  range_counts,
  color_sub = NULL,
  draw_points = TRUE,
  lwd = 1,
  ...
)
}
\arguments{
\item{color_sub}{\code{character} vector of R colors, named
by values in the \code{Category} column of \code{counts_df}.}

\item{draw_points}{\code{logical} indicating if individual points
should be drawn, or when \code{draw_points=FALSE} only the
line is drawn.}

\item{lwd}{\code{numeric} indicating the line width.}

\item{...}{additional arguments are ignored.}

\item{counts_df}{\code{data.frame} produced by
\code{peak_feature_ranged_counter()}.}
}
\value{
\code{ggplot} object
}
\description{
Plot feature-peak overlap counts by distance
}
\details{
This function is a small utility function intended
to display the individual counts produced by
\code{peak_feature_ranged_counter()}.
}
