% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pde-enrich.R
\name{plot_ranged_enrichment}
\alias{plot_ranged_enrichment}
\title{Plot ranged enrichment hypergeometric results}
\usage{
plot_ranged_enrichment(
  paramsDF,
  p_cutoff = 0.01,
  p_color = "grey40",
  draw_points = TRUE,
  lwd = 1,
  pt_cex = 3,
  color_sub = NULL,
  fill_sub = color_sub,
  facet_formula = Experiment ~ PeakSet,
  shape_sub = c(higher = 24, lower = 25),
  alpha_sub = c(Test = 1, Random = 0.075),
  y_include = c(0, 10),
  x_include = c(5),
  test_cat_title = "Category",
  PeakSet_title = "PeakSet",
  main = "Enrichment by distance from peaks",
  xlab = "Distance from peak center (kb)",
  ylab = "-log10(hypergeometric P-value)",
  scales = "free_y",
  ...
)
}
\arguments{
\item{paramsDF}{\code{data.frame} produced by
\code{peak_feature_ranged_enrichment()} with
\code{return_type="phyper"}.}

\item{p_cutoff}{\code{numeric} threshold used to draw a horizontal
line indicating a threshold for statistical significance.}

\item{p_color}{\code{character} with R color, used to color the
horizontal line defined by \code{p_cutoff}.}

\item{draw_points}{\code{logical} indicating whether each point
should be drawn, or when \code{draw_points=FALSE} only the
line is drawn.}

\item{lwd}{\code{numeric} indicating the line width.}

\item{pt_cex}{\code{numeric} indicating point size, used when
\code{draw_points=TRUE}.}

\item{color_sub}{\code{character} vector of R colors, whose names
correspond to category values tested for enrichment.}

\item{fill_sub}{\code{character} vector, same as \code{color_sub} but
used when a fill color is required.}

\item{facet_formula}{\code{formula} sufficient to inform
\code{facet_grid()} used to create separate ggplot panels.}

\item{shape_sub}{\code{integer} vector of R point shapes, whose
names are peak names tested for enrichment.}

\item{alpha_sub}{\code{numeric} vector of alpha transparency values
ranging between 0 and 1, whose names are \code{Test} and
\code{Random}. These values are only used when random permutation
tests have been performed, and the entries that resulted from
random permutations are stored in column \code{"Test"} and have
value \code{"Random"}.}

\item{y_include}{\code{numeric} vector used to extend the y-axis
range to include at least these values. This effect is useful
in requiring the y-axis includes y=0 for example, and y=10
so that the y-axis range spans at least a reasonable viewing
range for interpreting the data. The y-axis units are
\code{-log10(pvalue)} values.}

\item{x_include}{\code{numeric} vector used to extend the x-axis to
ensure the range includes this value.}

\item{test_cat_title}{\code{character} string used as the title for the
category legend.}

\item{PeakSet_title}{\code{character} string used as a legend for point
shape, only relevant when \code{draw_points=TRUE}.}

\item{main}{\code{character} string used as the title, called
with \code{ggtitle()}.}

\item{xlab, ylab}{\code{character} string used for the x-axis and y-axis
label, respectively.}

\item{scales}{\code{character} string passed to \code{facet_grid()} which
defines the y-axis and x-axis range across multiple facet
panels: \code{"free_y"} will allow each panel to have independent
y-axis range; \code{"fixed"} will use one y-axis range across
all facet panels.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{ggplot} object, which can be viewed by calling
\code{print()}, or further modified.
}
\description{
Plot ranged enrichment hypergeometric results
}
\details{
This function takes \code{"phyper"} output from
\code{peak_feature_ranged_enrichment()} and produces a plot
using \code{ggplot()}.
}
